+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add block_height to UTXOSpendReport for more efficient scans.
+
 ## [0.6.1] - 2026-01-09
 
 ### Fixed